requirements.txt
boto3==1.4.0
botocore==1.4.56
docutils==0.12
jmespath==0.9.0
python-dateutil==2.5.3
requests==2.11.1
s3transfer==0.1.4
six==1.10.0
twx.botapi==2.1.0
